FIXED: Loyalty feature combined with User due to lack of need for seperate logic
