+++ src/main/java/com/smart_shopping_list_expense_manager/java/smart_shopping_list_expense_manager/dto/UserDTO.java
@@ -1,5 +1,6 @@
 package com.smart_shopping_list_expense_manager.java.smart_shopping_list_expense_manager.dto;
 
+import com.smart_shopping_list_expense_manager.java.smart_shopping_list_expense_manager.enums.LoyaltyTierEnum;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -12,5 +13,7 @@ public class UserDTO {
     private UUID id;
     private String email;
     private String name;
+    private LoyaltyTierEnum loyaltyTier;
+    private Integer bonus_points;
     private List<ShoppingListDTO> shoppingLists;
 }
