--- src/main/java/com/smart_shopping_list_expense_manager/java/smart_shopping_list_expense_manager/repositories/LoyaltyRepository.java
@@ -1,12 +1,0 @@
-package com.smart_shopping_list_expense_manager.java.smart_shopping_list_expense_manager.repositories;
-
-import com.smart_shopping_list_expense_manager.java.smart_shopping_list_expense_manager.entities.LoyaltyEntity;
-import com.smart_shopping_list_expense_manager.java.smart_shopping_list_expense_manager.entities.UsersEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface LoyaltyRepository extends JpaRepository<LoyaltyEntity, UUID> {
-    Optional<LoyaltyEntity> findByUser(UsersEntity user);
-}
