--- src/main/java/com/smart_shopping_list_expense_manager/java/smart_shopping_list_expense_manager/services/ExampleService.java
@@ -1,4 +1,0 @@
-package com.smart_shopping_list_expense_manager.java.smart_shopping_list_expense_manager.services;
-
-public class ExampleService {
-}
